Move  ./pkg/opentelemetry-mapping-go/internal/sketchtest to ./pkg/util/quantile
