+++ pkg/opentelemetry-mapping-go/otlp/metrics/testhelper_test.go
@@ -15,8 +15,8 @@ import (
 	"testing"
 
 	"github.com/DataDog/datadog-agent/pkg/opentelemetry-mapping-go/otlp/attributes"
-	"github.com/DataDog/datadog-agent/pkg/opentelemetry-mapping-go/quantile"
-	"github.com/DataDog/datadog-agent/pkg/opentelemetry-mapping-go/quantile/summary"
+	"github.com/DataDog/datadog-agent/pkg/util/quantile"
+	"github.com/DataDog/datadog-agent/pkg/util/quantile/summary"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/component/componenttest"
