+++ pkg/util/quantile/go.mod
@@ -1,9 +1,9 @@
-module github.com/DataDog/datadog-agent/pkg/opentelemetry-mapping-go/quantile
+module github.com/DataDog/datadog-agent/pkg/util/quantile
 
 go 1.23.0
 
 require (
-	github.com/DataDog/datadog-agent/pkg/opentelemetry-mapping-go/internal/sketchtest v0.0.0-00010101000000-000000000000
+	github.com/DataDog/datadog-agent/pkg/util/quantile/sketchtest v0.0.0-00010101000000-000000000000
 	github.com/DataDog/sketches-go v1.4.7
 	github.com/dustin/go-humanize v1.0.1
 	github.com/stretchr/testify v1.10.0
@@ -16,6 +16,6 @@ require (
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
-replace github.com/DataDog/datadog-agent/pkg/opentelemetry-mapping-go/internal/sketchtest => ../../../pkg/opentelemetry-mapping-go/internal/sketchtest
+replace github.com/DataDog/datadog-agent/pkg/util/quantile/sketchtest => ../../../pkg/util/quantile/sketchtest
 
 retract v0.4.0 // see #107
