+++ pkg/util/quantile/sparse_test.go
@@ -10,7 +10,7 @@ import (
 	"math"
 	"testing"
 
-	"github.com/DataDog/datadog-agent/pkg/opentelemetry-mapping-go/quantile/summary"
+	"github.com/DataDog/datadog-agent/pkg/util/quantile/summary"
 
 	"github.com/stretchr/testify/require"
 )
